fix:add access control

+++ Backend/src/modules/students/student.controller.js
@@ -1,11 +1,21 @@
 import { Student } from "./student.model.js";
 import { Quiz } from "../quiz/quiz.model.js"
+import { Classroom } from "../classrooms/classroom.model.js";
 
 
 //get all students
 export const getAllStudents = async (req, res) => {
     try {
-        const students = await Student.find().populate('classroomId', 'name'); //  populate
+
+        //fetch classroom to check teacher
+        const classroom = await Classroom.find({teacherId: req.user.id});
+        if (!classroom) return res.status(404).json({ message: "Classroom not found" });
+
+        //extract  classsroomids
+        const classroomIds = classroom.map(c=>c._id);
+
+       //fetch students in those classroom
+        const students = await Student.find({classroomId:{$in:classroomIds}}).populate('classroomId', 'name'); //  populate
         if (!students || students.length === 0) {
             return res.status(404).json({ message: "students not found" });
         }
@@ -20,14 +30,23 @@ export const getAllStudents = async (req, res) => {
 //student insert
 export const addStudent = async(req, res) =>{
 
-    const {regNo, name, classroomId} = req.body;
+      const {regNo, name, classroomId} = req.body;
 
     
 
     try{
-    const student = await Student.create({regNo, name,classroomId});
-    return res.status(200).json({student});  
 
+        //fetch classroom to check teacher
+        const classroom = await Classroom.findById(classroomId);
+        if (!classroom) return res.status(404).json({ message: "Classroom not found" });
+
+        if (classroom.teacherId.toString() !== req.user.id) {
+            return res.status(403).json({ message: "Not authorized to add student to this classroom" });
+        }
+
+         const student = await Student.create({regNo, name,classroomId});
+         return res.status(200).json({student});  
+ 
     }catch(err){
         console.log(err);
         return res.status(500).json({ message: "Unable to add student" });
@@ -66,16 +85,26 @@ export const addStudent = async(req, res) =>{
           let student;
  
           try{
-             student = await Student.findByIdAndUpdate(id,
-                  { name },
-                  { returnDocument: 'after', runValidators: true }
-             );
+             student = await Student.findById(id);
+                 
               if (!student) return res.status(404).json({ message: "student not found" });
-         return res.status(200).json({ student});
+
+              //fetch classroom to check teacher
+                const classroom = await Classroom.findById(student.classroomId);
+                if (!classroom) return res.status(404).json({ message: "Classroom not found" });
+
+                 if (classroom.teacherId.toString() !== req.user.id) {
+                 return res.status(403).json({ message: "Not authorized to add student to this classroom" });
+                }
+
+                student.name = name || student.name;
+                await student.save();
+
+                return res.status(200).json({ student});
           }catch (err) {
-         console.error(err);
-         return res.status(500).json({ message: "Server error" });
-     }
+               console.error(err);
+               return res.status(500).json({ message: "Server error" });
+          }
 };
 
 //Delete student
@@ -84,9 +113,19 @@ export const addStudent = async(req, res) =>{
      
              let student;
              try{
-             student = await Student.findByIdAndDelete(id)
+             student = await Student.findById(id)
      
              if (!student) return res.status(404).json({ message: "student unable to delete" });
+
+             // check if logged teacher is assigned to this student's classroom
+              const classroom = await Classroom.findById(student.classroomId);
+              if (!classroom) return res.status(404).json({ message: "Classroom not found" });
+
+             if (classroom.teacherId.toString() !== req.user.id) {
+                return res.status(403).json({ message: "Not authorized to delete this student" });
+               }
+
+                await Student.findByIdAndDelete(id);
              return res.status(200).json({ student});
      
              }catch (err) {
@@ -103,7 +142,12 @@ export const addStudent = async(req, res) =>{
      export const getActiveQuizForStudent = async(req,res) =>{
               
             try{
-                 const classroomId = "699dd1ae870cb401c2741497";
+
+                const studentId = req.user.id; // from authMiddleware
+                const student = await Student.findById(studentId);
+                if (!student) return res.status(404).json({ message: "Student not found" });
+                
+                const classroomId = student.classroomId;//get the id from populated
                  const now = new Date();
 
                  const quiz = await Quiz.findOne({
